refactor: move example_app, example_util to packages: demo, strings
