+++ apps/example_app/src/main/java/org/sab/demo/App.java
@@ -1,4 +1,6 @@
-package org.sab;
+package org.sab.demo;
+
+import org.sab.strings.ExampleUtil;
 
 import java.util.List;
