+++ apps/example_app/src/test/java/org/sab/demo/AppTest.java
@@ -1,4 +1,4 @@
-package org.sab;
+package org.sab.demo;
 
 import static org.junit.Assert.assertTrue;
